AUT-3924: Extract TXMA header method
This will allow it to be reused more easily outside of the base frontend handler
